+++ src/ui_helper.rs
@@ -1,108 +1,98 @@
-use pbr::{MultiBar, Pipe, ProgressBar, Units};
-use std::io::Stdout;
+use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
 use std::sync::Mutex;
 use std::thread;
-use std::time::Duration;
 
 lazy_static! {
     static ref TOTALS: Mutex<Vec<u64>> = Mutex::new(vec![]);
-    static ref PBRS: Mutex<Vec<ProgressBar<Pipe>>> = Mutex::new(vec![]);
+    static ref PBRS: Mutex<Vec<ProgressBar>> = Mutex::new(vec![]);
 }
 
-pub fn start_pbr(file_name: &str, lengths: Vec<u64>) {
-    let mut mb = MultiBar::new();
-    mb.println(&format!("Downloading: {}", file_name));
+pub fn start_all_pb(file_name: &str, thread_count: usize, lengths: Vec<u64>) {
+    let mut mb = MultiProgress::new();
+    mb.set_move_cursor(true);
+
+    println!("{}", file_name);
 
     let total_length = lengths.iter().sum();
-    build_global_bar(&mut mb, total_length);
-    mb.println("");
+    let master_style = ProgressStyle::default_bar()
+        .template("[{elapsed_precise}] {percent}% [{wide_bar:.green}] {bytes}/{total_bytes}")
+        .progress_chars("█▉▊▋▌▍▎▏  ");
+    build_bar(&mut mb, master_style, total_length);
 
-    for length in lengths {
-        build_child_bar(&mut mb, length);
+    {
+        let mut totals = TOTALS
+            .lock()
+            .expect("Failed to acquire TOTALS lock, lock poisoned!");
+        for _i in 0..lengths.len() {
+            totals.push(0);
+        }
     }
 
-    thread::spawn(move || mb.listen());
+    for &length in lengths.iter().take(thread_count) {
+        let child_style = ProgressStyle::default_bar()
+            .template("{msg} [{wide_bar:.yellow}] {bytes}/{total_bytes}")
+            .progress_chars("█▉▊▋▌▍▎▏  ");
+
+        build_bar(&mut mb, child_style.clone(), length);
+    }
+
+    thread::spawn(move || mb.join_and_clear().unwrap());
 }
 
-pub fn setting_up_bar(bar_idx: usize) {
-    let mut pbrs = PBRS
+pub fn setting_up_bar(thread_id: usize, part_id: usize, size: u64) {
+    let pbrs = PBRS
         .lock()
         .expect("Failed to acquire PBRS lock, lock poisoned!");
-    pbrs[bar_idx + 1].message("Starting... ");
-    pbrs[bar_idx + 1].tick();
+
+    pbrs[thread_id].set_length(size);
+    pbrs[thread_id].set_position(0);
+    pbrs[thread_id].set_message(&format!("Part {}", part_id + 1));
+    pbrs[thread_id].tick();
 }
 
-pub fn start_bar(bar_idx: usize) {
-    let mut pbrs = PBRS
-        .lock()
-        .expect("Failed to acquire PBRS lock, lock poisoned!");
-    pbrs[bar_idx + 1].message("");
-    pbrs[bar_idx + 1].show_message = false;
-    pbrs[bar_idx + 1].tick();
-}
-
-pub fn update_bar(bar_idx: usize, progress: u64) {
-    let mut pbrs = PBRS
-        .lock()
-        .expect("Failed to acquire PBRS lock, lock poisoned!");
+pub fn adjust_totals(part_id: usize, progress: u64) {
     let mut totals = TOTALS
         .lock()
         .expect("Failed to acquire TOTALS lock, lock poisoned!");
 
-    pbrs[bar_idx + 1].set(progress);
-    totals[bar_idx] = progress;
+    totals[part_id] = progress;
 
     let total_progress = totals.iter().sum();
-    pbrs[0].set(total_progress);
-}
-
-pub fn success_global_bar() {
-    finish_bar_with_message(0, "Download Complete!");
-}
-
-pub fn success_bar(bar_idx: usize) {
-    finish_bar_with_message(bar_idx + 1, "Download Complete!");
-}
-
-pub fn fail_bar(bar_idx: usize) {
-    finish_bar_with_message(bar_idx + 1, "Download Failed!");
-}
-
-fn finish_bar_with_message(act_bar: usize, message: &str) {
     PBRS.lock()
-        .expect("Failed to acquire PBRS lock, lock poisoned!")[act_bar]
-        .finish_print(message);
+        .expect("Failed to acquire PBRS lock, lock poisoned!")[0]
+        .set_position(total_progress);
 }
 
-fn build_global_bar(mb: &mut MultiBar<Stdout>, size: u64) {
-    build_bar(mb, size, None);
+pub fn update_bar(thread_id: usize, part_id: usize, progress: u64) {
+    {
+        PBRS.lock()
+            .expect("Failed to acquire PBRS lock, lock poisoned!")[thread_id]
+            .set_position(progress);
+    }
+
+    adjust_totals(part_id, progress);
 }
 
-fn build_child_bar(mb: &mut MultiBar<Stdout>, size: u64) {
-    build_bar(mb, size, Some("Pending... ".to_string()));
-
-    let mut totals = TOTALS
-        .lock()
-        .expect("Failed to acquire TOTALS lock, lock poisoned!");
-    totals.push(0);
+pub fn success_bar(thread_id: usize) {
+    finish_bar_with_message(thread_id, "Done");
 }
 
-fn build_bar(mb: &mut MultiBar<Stdout>, size: u64, message: Option<String>) {
+pub fn fail_bar(thread_id: usize) {
+    finish_bar_with_message(thread_id, "Download Failed!");
+}
+
+fn finish_bar_with_message(thread_id: usize, message: &str) {
+    PBRS.lock()
+        .expect("Failed to acquire PBRS lock, lock poisoned!")[thread_id]
+        .finish_with_message(message);
+}
+
+fn build_bar(mb: &mut MultiProgress, style: ProgressStyle, size: u64) {
     let mut pbrs = PBRS
         .lock()
         .expect("Failed to acquire PBRS lock, lock poisoned!");
-    let mut pb = mb.create_bar(size);
-    pb.set_max_refresh_rate(Some(Duration::from_millis(200)));
-    pb.tick_format("▏▎▍▌▋▊▉██▉▊▋▌▍▎▏");
-    pb.set_units(Units::Bytes);
-
-    if let Some(msg) = message {
-        pb.show_message = true;
-        pb.message(&msg);
-    } else {
-        pb.show_message = false;
-    }
-
+    let pb = mb.add(ProgressBar::new(size));
+    pb.set_style(style);
     pb.tick();
     pbrs.push(pb);
 }